Weapon Model Swap
Got the revolver working, need to do the shotgun. Also need to change where the pickup spawner locates the weapon because as of now it is using the default unreal pistol which I swapped models for testing. Need to change it to the pistol that weapons that Nathan is working on.